Enhancement minerva update (#120)
* fixing issues with icd parsing

* Update for minerva

* Big update for minerva

* Moved mmeds.py to util.py

* Fixed core_diversity_analysis

* split util.py

* should be good now
