+++ .travis.yml
@@ -3,61 +3,62 @@ python: '3.7'
 dist: xenial
 sudo: true
 addons:
-    apt:
-        packages:
-            - texlive-latex-base
-            - texlive-latex-extra
-            - texlive-latex-recommended
-            - texlive-generic-recommended
-            - texlive-pictures
-            - texlive-publishers
-            - texlive-fonts-recommended
-            - texlive-fonts-extra
-            - cmake
-            - xsltproc
+  apt:
+    packages:
+    - texlive-latex-base
+    - texlive-latex-extra
+    - texlive-latex-recommended
+    - texlive-generic-recommended
+    - texlive-pictures
+    - texlive-publishers
+    - texlive-fonts-recommended
+    - texlive-fonts-extra
+    - cmake
+    - xsltproc
 before_cache:
-    - sudo chown -R travis:travis $HOME
-    - sudo chown -R travis:travis $TRAVIS_BUILD_DIR
+- sudo chown -R travis:travis $HOME
+- sudo chown -R travis:travis $TRAVIS_BUILD_DIR
 cache:
-    directories:
-        - ~/.modules
-        - ~/.local
-    timeout: 1000
-    apt: true
+  directories:
+  - "~/.modules"
+  - "~/.local"
+  timeout: 1000
+  apt: true
 git:
-    depth: 1
+  depth: 1
 services:
-    - mysql
-    - mongodb
-    - xvfb
+- mysql
+- mongodb
+- xvfb
 before_install:
-    - openssl aes-256-cbc -K $encrypted_754fbe69cba7_key -iv $encrypted_754fbe69cba7_iv
-      -in ./mmeds/resources/privkey.pem.enc -out ./mmeds/resources/privkey.pem -d
-    - openssl aes-256-cbc -K $encrypted_ad273e68e13c_key -iv $encrypted_ad273e68e13c_iv
-      -in ./mmeds/secrets.py.enc -out ./mmeds/secrets.py -d
-    - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O ~/miniconda.sh --quiet
-    - chmod +x ~/miniconda.sh
-    - ~/miniconda.sh -b &>/dev/null
-    - export PATH=~/miniconda2/bin:$PATH
-    - conda update --yes conda &>/dev/null
-    - sudo chown -R travis:travis ~/.modules/modulefiles
-    - sudo chown -R travis:travis ~/.modules ~/.local
-    - bash travis_setup.sh
-    - source ~/.bashrc
-    - conda info --envs
+- openssl aes-256-cbc -K $encrypted_0fabcfc21ca1_key -iv $encrypted_0fabcfc21ca1_iv
+  -in ./mmeds/secrets.py.enc -out ./mmeds/secrets.py -d
+- openssl aes-256-cbc -K $encrypted_754fbe69cba7_key -iv $encrypted_754fbe69cba7_iv
+  -in ./mmeds/resources/privkey.pem.enc -out ./mmeds/resources/privkey.pem -d
+- wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O ~/miniconda.sh
+  --quiet
+- chmod +x ~/miniconda.sh
+- "~/miniconda.sh -b &>/dev/null"
+- export PATH=~/miniconda2/bin:$PATH
+- conda update --yes conda &>/dev/null
+- sudo chown -R travis:travis ~/.modules/modulefiles
+- sudo chown -R travis:travis ~/.modules ~/.local
+- bash travis_setup.sh
+- source ~/.bashrc
+- conda info --envs
 install:
-    - source activate mmeds-stable
-    - pip install . --no-cache-dir
-    - mysql -u root < setup.sql
+- source activate mmeds-stable
+- pip install . --no-cache-dir
+- mysql -u root < setup.sql
+- ls ~/.modules/modulefiles
+- cat ~/.modules/modulefiles/qiime2/2019.1
 script:
-    - pytest --cov=mmeds -W ignore::DeprecationWarning -W ignore::FutureWarning -x
-    # - cd server; timeout 333 python server.py 1&>/dev/null& sleep 1; locust -c 5 -r 1 -t 5m --no-web --only-summary -f ../mmeds/tests/server_stress.py
-    - tail -n 50 ~/mmeds_server_data/mmeds_log.txt
-    # Clear modified directories so travis doesn't have to update the cache
-    - pip uninstall mmeds -y
-    - sudo rm -rf ~/.modules/qiime2/var/cache
-    - sudo rm -rf ~/.modules/qiime1/var/cache
-    - sudo rm -rf ~/.modules/mmeds-stable/var/cache
-    - sudo find ~/.modules/ -type f -name *.pyc -delete
+- travis_wait 30 pytest --cov=mmeds -W ignore::DeprecationWarning -W ignore::FutureWarning -x
+- tail -n 50 ~/mmeds_server_data/mmeds_log.txt
+- pip uninstall mmeds -y
+- sudo rm -rf ~/.modules/qiime2/var/cache
+- sudo rm -rf ~/.modules/qiime1/var/cache
+- sudo rm -rf ~/.modules/mmeds-stable/var/cache
+- sudo find ~/.modules/ -type f -name *.pyc -delete
 after_success:
-    - bash <(curl -s https://codecov.io/bash)
+- bash <(curl -s https://codecov.io/bash)
