+++ README.md
@@ -1,8 +1,8 @@
 # mmeds-meta
 
-[![Build Status](https://travis-ci.com/clemente-lab/mmeds-meta.png?branch=Enhancement-BetterSummaryFile)](https://travis-cl.com/clemente-lab/mmeds-meta)
-[![codecov](https://codecov.io/gh/clemente-lab/mmeds-meta/branch/Enhancement-BetterSummaryFile/graph/badge.svg)](https://codecov.io/gh/clemente-lab/mmeds-meta)
-[![CodeFactor](https://codefactor.io/repository/github/clemente-lab/mmeds-meta/badge/Enhancement-BetterSummaryFile)](https://codefactor.io/repository/github/clemente-lab/mmeds-meta/overview/Enhancement-BetterSummaryFile)
+[![Build Status](https://travis-ci.com/clemente-lab/mmeds-meta.png?branch=Enhancement-MinervaUpdate)](https://travis-cl.com/clemente-lab/mmeds-meta)
+[![codecov](https://codecov.io/gh/clemente-lab/mmeds-meta/branch/Enhancement-MinervaUpdate/graph/badge.svg)](https://codecov.io/gh/clemente-lab/mmeds-meta)
+[![CodeFactor](https://codefactor.io/repository/github/clemente-lab/mmeds-meta/badge/Enhancement-MinervaUpdate)](https://codefactor.io/repository/github/clemente-lab/mmeds-meta/overview/Enhancement-MinervaUpdate)
