+++ mmeds/resources/summary_code.txt
@@ -8,12 +8,12 @@ from IPython.display import Image
 from random import shuffle
 from PIL import Image as PImage
 from PIL import ImageDraw, ImageFont
-from mmeds.mmeds import load_config
+from mmeds.util import load_config
 
 filterwarnings('ignore', category=rpy2.rinterface.RRuntimeWarning)
 
 # Load the configuration
-config = load_config(Path('../config_file.txt').read_text(), Path('../metadata.tsv'))
+config = load_config(Path('config_file.txt').read_text(), Path('metadata.tsv'))
 
 # Set parameters used when making the legend
 font_size = int(config['font_size'])
