+++ mmeds/test_files/User_Guide.txt
@@ -49,38 +49,30 @@ Result: The final result of running a computational tool (such as Qiime)
         on RawData. This will typically be generated by MMEDS and so it
         may be left blank.
 
+
+For all fields entries should contain no special characters '$', '*', ';', etc.
+Metadata files containing entries with these characters cannot be uploaded to the server.
+All text fields will be converted to lowercase for storage in the database
+so be sure none of the entries are case sensitive.
+
 Certain Fields will only accept values that fall into certain categories.
 They are as follows:
 
+RawDataID:
+    - This is the unique identifier for each sample that has already gone through sequencing
+    - Equivalent to #SampleID when using qiime
+    - If the reads are demultiplexed the file names should follow the format
+        - {RawDataID}_{AdditionalInformation}_R1_001.fastq.gz
+        - E.G.
+            - Forward Read: S33_L001_R1_001.fastq.gz
+            - Reverse Read: S33_L001_R2_001.fastq.gz
+SpecimenBodySite & SpecimenType:
+https://www.ebi.ac.uk/ols/ontologies/uberon
+    - Values should be taken from UBERON, using the terms as provided on the website
+    - The UBERON ID should be included in their respective UberonCode columns
+Ethnicity:
 Categories taken from NIH standards.
 https://grants.nih.gov/grants/guide/notice-files/not-od-15-089.html
-SpecimenBodySite:
-    - Head
-    - Nose
-    - Ear
-    - Neck
-    - Hand
-    - Arm
-    - Torso-Front
-    - Torso-back
-    - Pelvic-Front
-    - Pelvic-Back
-    - Genital
-    - Rectal
-    - Leg
-    - Foot
-    - Other
-SpecimenType:
-    - Skin
-    - Oral
-    - Nasal
-    - Fecal
-    - Gut
-    - Ear
-    - Saliva
-    - Internal Tissue
-    - Other
-Ethnicity:
     - American Indian or Alaska Native
     - Asian
     - Black or African American
