+++ mmeds/test_files/test_config_file.txt
@@ -5,7 +5,7 @@ sampling_depth	1114
 # Otherwise list all the column names separated by commas
 # Example:
 # metadata	all
-metadata	Ethnicity,DaysSinceExperimentStart
+metadata	Ethnicity
 # The levels of taxa summary to include in the final summary of the analysis
 taxa_levels	2
 # The minimum abundance a taxon must have in a sample in order to be included in the legend
