+++ mmeds/util.py
@@ -1,33 +1,13 @@
 from collections import defaultdict
-from numpy import std, mean, issubdtype, number, nan
-from numpy import datetime64
-from mmeds.error import MetaDataError, InvalidConfigError, InvalidSQLError, InvalidModuleError
+from mmeds.error import InvalidConfigError, InvalidSQLError, InvalidModuleError
 from subprocess import run
 from datetime import datetime
 from pathlib import Path
 from os import environ
+from numpy import nan, issubdtype, int64, float64, datetime64, number
+
 import mmeds.config as fig
 import pandas as pd
-import numpy as np
-
-NAs = ['n/a', 'n.a.', 'n_a', 'na', 'N/A', 'N.A.', 'N_A']
-
-REQUIRED_HEADERS = set(['Description',
-                        '#SampleID'
-                        'BarcodeSequence'
-                        'LinkerPrimerSequence'
-                        'Lab'
-                        'AnalysisTool'
-                        'PrimaryInvestigator'])
-
-REQUIRED_HEADERS = set(['SpecimenID', 'BarcodeSequence', 'PrimaryInvestigator'])
-
-HIPAA_HEADERS = ['social_security', 'social_security_number', 'address', 'phone', 'phone_number']
-
-DNA = set('GATC')
-
-ILLEGAL_IN_HEADER = set('/\\ *?')  # Limit to alpha numeric, underscore, dot, hyphen, has to start with alpha
-ILLEGAL_IN_CELL = set(str(ILLEGAL_IN_HEADER) + '_')
 
 
 def load_config(config_file, metadata):
@@ -133,9 +113,62 @@ def get_valid_columns(metadata_file, option):
     return summary_cols, col_types
 
 
+def get_col_type(raw_column):
+    """
+    Return the type of data the column should be checked for.
+    =========================================================
+    :raw_column: The column to check for type
+    """
+    check_date = False
+    col_type = None
+    if 'Date' in raw_column.name:
+        col_type = datetime64
+        try:
+            column = pd.to_datetime(raw_column)
+        # If there is an error converting to datetime
+        # check the individual cells
+        except ValueError:
+            column = raw_column
+            check_date = True
+    # Try to set the type based on the most common type
+    else:
+        column = raw_column
+        types = {
+            int: 0,
+            float: 0,
+            str: 0
+        }
+
+        for cell in raw_column:
+            log('cell: {}, is_num: {}'.format(cell, is_numeric(cell)))
+            # Don't count NA
+            if cell == 'NA':
+                continue
+            # Check if value is numeric
+            elif is_numeric(cell):
+                try:
+                    int(cell)
+                    types[int] += 1
+                except ValueError:
+                    types[float] += 1
+            # Check if it's a string
+            else:
+                try:
+                    str(cell)
+                    types[str] += 1
+                except TypeError:
+                    continue
+        col_type = max(types, key=types.get)
+    return column, col_type, check_date
+
+
 def load_ICD_codes():
     """ Load all known ICD codes and return them as a dictionary """
-    ICD_codes = {'XXX.XXXX': 'Subject is healthy to the best of our knowledge'}
+    ICD_codes = {
+        'XXX.XXXX': 'Subject is healthy to the best of our knowledge',
+        'NA': 'No Value',
+        nan: 'No Value'
+    }
     with open(fig.STORAGE_DIR / 'icd10cm_codes_2018.txt') as f:
         # Parse each line
         for line in f:
@@ -150,6 +183,88 @@ def load_ICD_codes():
             code = code[:3] + '.' + code[3:]
             ICD_codes[code] = description
     return ICD_codes
+
+
+def load_mapping_file(file_fp, delimiter):
+    """
+    Load the metadata file and assign datatypes to the columns
+    ==========================================================
+    :file_fp: The path to the mapping file
+    :delimiter: The delimiter used in the mapping file
+    """
+    df = pd.read_csv(file_fp,
+                     sep=delimiter,
+                     header=[0, 1],
+                     skiprows=[2, 3, 4],
+                     na_filter=False)
+    df.replace('NA', nan, inplace=True)
+    # Get the tables in the dataframe while maintaining order
+    tables = []
+    errors = []
+    warnings = []
+    for (table, header) in df.axes[1]:
+        tables.append(table)
+        for column in df[table]:
+            log(df[table][column])
+            if '' in df[table][column]:
+                errors.append('-1\t-1\tColumn Value Error: Column {} is missing entries'.format(column))
+                continue
+            try:
+                df[table].assign(column=df[table][column].astype(fig.COLUMN_TYPES[table][column]))
+            # Additional metadata won't have an entry so will automatically be treated as a string
+            except KeyError:
+                df[table].assign(column=df[table][column].astype('object'))
+            # Error handling for column values that don't match the column type
+            except ValueError:
+                errors.append('-1\t-1\tColumn Value Error: Column {} contains the wrong type of values'.format(column))
+    tables = list(dict.fromkeys(tables))
+    return tables, df, errors, warnings
+
+
+def load_html(file_path, **kwargs):
+    """
+    Load the specified html file. Inserting the head and topbar
+    """
+    # Load the html page
+    with open(fig.HTML_DIR / (file_path + '.html')) as f:
+        page = f.read().split('\n')
+
+    # Load the head information
+    with open(fig.HTML_DIR / 'header.html') as f:
+        header = f.read().split('\n')
+
+    # Load the topbar information
+    with open(fig.HTML_DIR / 'topbar.html') as f:
+        topbar = f.read().split('\n')
+
+    new_page = page[:3] + header + topbar + page[3:]
+    return '\n'.join(new_page).format(**kwargs)
+
+
+def load_MIxS_metadata(file_fp, skip_rows, unit_column):
+    """
+    A function for load and transforming the MIxS data in a pandas dataframe.
+    ========================================================================
+    :file_fp: The path to the file to convert
+    :skip_rows: The number of rows to skip after the header
+    :unit_column: A string. If None then the function checks each cell for units.
+    """
+    units = {}
+    # Read in the data file
+    df = pd.read_csv(file_fp, header=0, sep='\t')
+    # Set the index to be the 'column_header' column
+    df.set_index('column_header', inplace=True)
+    # Remove rows with null indexes
+    df = df.loc[df.index.notnull()]
+    # Transpose the dataframe across the diagonal
+    df = df.T
+    # Drop unnamed columns
+    df.drop([x for x in df.axes[0] if 'Unnamed' in x], inplace=True)
+    # Drop any columns with only np.nan values
+    df.dropna(how='all', axis='columns', inplace=True)
+    # Replace np.nans with "NA"s
+    df.fillna('"NA"', inplace=True)
+    return df, units
 
 
 def insert_error(page, line_number, error_message):
@@ -184,436 +299,6 @@ def insert_html(page, line_number, html):
     return new_page
 
 
-def load_html(file_path, **kwargs):
-    """
-    Load the specified html file. Inserting the head and topbar
-    """
-    # Load the html page
-    with open(fig.HTML_DIR / (file_path + '.html')) as f:
-        page = f.read().split('\n')
-
-    # Load the head information
-    with open(fig.HTML_DIR / 'header.html') as f:
-        header = f.read().split('\n')
-
-    # Load the topbar information
-    with open(fig.HTML_DIR / 'topbar.html') as f:
-        topbar = f.read().split('\n')
-
-    new_page = page[:2] + header + topbar + page[2:]
-    return '\n'.join(new_page).format(**kwargs)
-
-
-def check_header(header, col_index):
-    """ Check the header field to ensure it complies with MMEDS requirements. """
-    errors = []
-
-    row_col = '0\t' + str(col_index) + '\t'
-
-    # Check if it's numeric
-    if is_numeric(header):
-        text = 'Number Header Error: Column names cannot be numbers. Replace header %s of column\t%d '
-        errors.append(row_col + text % (header, col_index))
-    # Check if it's NA
-    if header in NAs + ['NA']:
-        errors.append(row_col + 'NA Header Error: Column names cannot be NA. Replace  header %s of column\t%d ' %
-                      (header, col_index))
-    # Check for illegal characters
-    if ILLEGAL_IN_HEADER.intersection(set(header)):
-        illegal_chars = ILLEGAL_IN_HEADER.intersection(set(header))
-        errors.append(row_col + 'Illegal Header Error: Illegal character(s) %s. Replace header %s of column\t%d' %
-                      (' '.join(illegal_chars), header, col_index))
-    # Check for HIPAA non-compliant headers
-    if header.lower() in HIPAA_HEADERS:
-        errors.append(row_col + 'PHI Header Error: Potentially identifying information in %s of column\t%d' %
-                      (header, col_index))
-    # Check for trailing or preceding whitespace
-    if not header == header.strip():
-        errors.append(row_col + 'Whitespace Header Error: Preceding or trailing whitespace %s in column %d' %
-                      (header, col_index))
-    return errors
-
-
-def check_cell(row_index, col_index, cell, col_type, check_date):
-    """
-    Check the data in the specified cell.
-    ====================================
-    :row_index: The index of the row of the cell
-    :col_index: The index of the column of the cell
-    :cell: The value of the cell
-    :col_type: The known type of the column as a whole
-    :check_date: If True check the cell for a valid date
-    """
-    # An NA cell will not generate any errors
-    if cell == 'NA':
-        return []
-    errors = []
-    row_col = str(row_index) + '\t' + str(col_index) + '\t'
-    # Check for non-standard NAs
-    if cell in NAs:
-        errors.append(row_col + 'NA Error: Non standard NA format %s\t%d,%d' %
-                      (cell, row_index, col_index))
-
-    # Check for consistent types in the column
-    if not (issubdtype(col_type, datetime64) or issubdtype(col_type, object)):
-        # If the cast fails for this cell the data must be the wrong type
-        try:
-            col_type(cell)
-        except ValueError:
-            errors.append(row_col + 'Mixed Type Error: Value {} does not match column type {}'.format(cell, col_type))
-    # Check for empty fields
-    if '' == cell:
-        errors.append(row_col + 'Empty Cell Error: Empty cell value %s' % cell)
-
-    if isinstance(cell, str):
-        # Check for trailing or preceding whitespace
-        if not cell == cell.strip():
-            errors.append(row_col + 'Whitespace Error: Preceding or trailing whitespace %s in row %d' %
-                          (cell, row_index))
-    # Check if this is the cell with the invalid date
-    if check_date:
-        try:
-            pd.to_datetime(cell)
-        except ValueError:
-            errors.append(row_col + 'Date Error: Invalid date {} in row {}'.format(cell, row_index))
-    return errors
-
-
-def get_col_type(raw_column):
-    """
-    Return the type of data the column should be checked for.
-    =========================================================
-    :raw_column: The column to check for type
-    """
-    check_date = False
-    col_type = None
-    if 'Date' in raw_column.name:
-        col_type = datetime64
-        try:
-            column = pd.to_datetime(raw_column)
-        # If there is an error converting to datetime
-        # check the individual cells
-        except ValueError:
-            column = raw_column
-            check_date = True
-    # Try to set the type based on the most common type
-    else:
-        column = raw_column
-        types = {
-            int: 0,
-            float: 0,
-            str: 0
-        }
-
-        for cell in raw_column:
-            # Don't count NA
-            if cell == 'NA':
-                continue
-            # Check if value is numeric
-            elif is_numeric(cell):
-                try:
-                    int(cell)
-                    types[int] += 1
-                except ValueError:
-                    types[float] += 1
-            # Check if it's a string
-            else:
-                try:
-                    str(cell)
-                    types[str] += 1
-                except TypeError:
-                    continue
-        col_type = max(types, key=types.get)
-    return column, col_type, check_date
-
-
-def check_number_column(column, col_index, col_type):
-    """ Check for mixed types and values outside two standard deviations. """
-    warnings = []
-    try:
-        filtered = [col_type(x) for x in column.tolist() if not x == 'NA']
-        stddev = std(filtered)
-        avg = mean(filtered)
-        for i, cell in enumerate(column):
-            if not cell == 'NA' and (col_type(cell) > avg + (2 * stddev) or col_type(cell) < avg - (2 * stddev)):
-                text = '%d\t%d\tStdDev Warning: Value %s outside of two standard deviations of mean in column %d'
-                warnings.append(text % (i + 1, col_index, cell, col_index))
-    except ValueError:
-        warnings.append("-1\t-1\tMixed Type Warning: Cannot get average of column {}. Mixed types".format(column))
-    return warnings
-
-
-def check_string_column(column, col_index):
-    """ Check for categorical data. """
-    warnings = []
-    counts = column.value_counts()
-    stddev = std(counts.values)
-    avg = mean(counts.values)
-    for val, count in counts.iteritems():
-        if count < (avg - stddev) and count < 3:
-            text = '%d\t%d\tCategorical Data Warning: Potential categorical data detected.\
-                Value %s may be in error, only %d found.'
-            warnings.append(text % (-1, col_index, val, count))
-    return warnings
-
-
-def check_column(raw_column, col_index):
-    """
-    Validate that there are no issues with the provided column of metadata.
-    =======================================================================
-    :raw_column: The unmodified column from the metadata dataframe
-    :col_index: The index of the column in the original dataframe
-    """
-    column, col_type, check_date = get_col_type(raw_column)
-
-    # Get the header
-    header = column.name
-
-    # Check the header
-    errors = check_header(header, col_index)
-    warnings = []
-
-    # Check the remaining columns
-    for i, cell in enumerate(column):
-        errors += check_cell(i, col_index, cell, col_type, check_date)
-
-    # Ensure there is only one study being uploaded
-    if header == 'StudyName' and len(set(column.tolist())) > 1:
-        errors.append('-1\t-1\tMultiple Studies Error: Multiple studies in one metadata file')
-
-    # Check that values fall within standard deviation
-    if issubdtype(col_type, number) and not isinstance(raw_column.dtype, object):
-        warnings += check_number_column(column, col_index, col_type)
-    # Check for categorical data
-    elif issubdtype(col_type, str) and not header == 'ICDCode':
-        warnings += check_string_column(column, col_index)
-
-    return errors, warnings
-
-
-def check_duplicates(column, col_index):
-    """ Checks for any duplicate entries in the provided column """
-
-    errors = []
-    cells = defaultdict(list)
-
-    # Add the indices of each item
-    for i, cell in enumerate(column):
-        cells[cell].append(i)
-    # Find any duplicates
-    dups = {k: v for k, v in cells.items() if len(v) > 1}
-    for dup_key in dups.keys():
-        value = dups[dup_key]
-        for val in value[1:]:
-            errors.append('%d\t%d\tDuplicate Value Error: Duplicate value of row %d, %s in row %d.' %
-                          (val, col_index, value[0], dup_key, val))
-    return errors
-
-
-def check_lengths(column, col_index):
-    """ Checks that all entries have the same length in the provided column """
-    errors = []
-    length = len(column[0])
-    for i, cell in enumerate(column[1:]):
-        if not len(cell) == length:
-            errors.append('%d\t%d\tLength Error: Value %s has a different length from other values in column %d' %
-                          (i + 1, col_index, cell, col_index))
-    return errors
-
-
-def check_barcode_chars(column, col_index):
-    """ Check that BarcodeSequence only contains valid DNA characters. """
-    errors = []
-    for i, cell in enumerate(column):
-        diff = set(cell).difference(DNA)
-        if diff:
-            errors.append('%d\t%d\tBarcode Error: Invalid BarcodeSequence char(s) %s in row %d' %
-                          (i, col_index, ', '.join(diff), i))
-    return errors
-
-
-def check_duplicate_cols(headers):
-    """
-    Returns true if there are any duplicate headers.
-    ================================================
-    :headers: The headers for each column in the metadata file
-    """
-    dups = []
-    for header in headers:
-        if '.1' in header:
-            dups.append(header.split('.')[0])
-    return dups
-
-
-def check_dates(df):
-    """
-    Check that no dates in the end col are earlier than
-    the matching date in the start col
-    ===================================================
-    :df: The data frame of the table containing the columns
-    :table_col: The column of the offending start date:w
-
-    """
-    errors = []
-    start_col = 0
-    for i in range(len(df)):
-        if df['StartDate'][i] > df['EndDate'][i]:
-            err = '{}\t{}\tData Range Error: End date {} is earlier than start date {} in row {}'
-            errors.append(err.format(i + 1, start_col, df['EndDate'][i], df['StartDate'][i], i))
-    return errors
-
-
-def check_ICD_codes(column, col_index):
-    """ Ensures all ICD codes in the column are valid. """
-    # Load the ICD codes
-    ICD_codes = load_ICD_codes()
-    errors = []
-    for i, cell in enumerate(column):
-        if ICD_codes.get(cell) is None:
-            errors.append('{}\t{}\tICD Code Error: Invalid ICD code {} in row {}'.format(i, col_index, cell, i))
-    return errors
-
-
-def check_table_column(table_df, name, header, col_index, row_index, study_name):
-    errors = []
-    warnings = []
-    if not name == 'AdditionalMetaData' and header not in fig.TABLE_COLS[name]:
-        errors.append('-1\t{}\tColumn Table Error: Column {} should not be in table {}'.format(col_index, header, name))
-    col = table_df[header]
-    new_errors, new_warnings = check_column(col, col_index)
-    errors += new_errors
-    warnings += new_warnings
-
-    # Perform column specific checks
-    if name == 'Specimen':
-        if header == 'BarcodeSequence':
-            errors += check_duplicates(col, col_index)
-            errors += check_lengths(col, col_index)
-            errors += check_barcode_chars(col, col_index)
-        elif header == 'RawDataID':
-            errors += check_duplicates(col, col_index)
-        elif header == 'LinkerPrimerSequence':
-            errors += check_lengths(col, col_index)
-    elif name == 'ICDCode':
-        errors += check_ICD_codes(col, col_index)
-    elif study_name is None and name == 'Study':
-        study_name = table_df['StudyName'][row_index]
-    return errors, warnings
-
-
-def check_table(table_df, name, all_headers, study_name):
-    """
-    Check the data within a particular table
-    ========================================
-    :table_df: A pandas dataframe containing the data for the specified table
-    :name: The name of the table
-    :all_headers: The headers that have been encountered so far
-    :study_name: None if no StudyName column has been seen yet,
-        otherwise with have the previously seen StudyName
-    """
-    errors = []
-    warnings = []
-    start_col = None
-    end_col = None
-    if not name == 'AdditionalMetaData':
-        missing_cols = set(fig.TABLE_COLS[name]).difference(table_df.columns)
-        if missing_cols:
-            text = '-1\t-1\tMissing Column Error: Columns {} missing from table {}'
-            errors.append(text.format(', '.join(missing_cols), name))
-    # For each table column
-    for i, header in enumerate(table_df.columns):
-        # Check that end dates are after start dates
-        if header == 'StartDate':
-            start_col = i
-        elif header == 'EndDate':
-            end_col = i
-        col_index = len(all_headers)
-        new_errors, new_warnings = check_table_column(table_df,
-                                                      name,
-                                                      header,
-                                                      col_index,
-                                                      i,
-                                                      study_name)
-        all_headers.append(header)
-        errors += new_errors
-        warnings += new_warnings
-    # Compare the start and end dates
-    if start_col is not None and end_col is not None:
-        errors += check_dates(table_df)
-    return (errors, warnings, all_headers, study_name)
-
-
-def load_mapping_file(file_fp, delimiter):
-    """
-    Load the metadata file and assign datatypes to the columns
-    ==========================================================
-    :file_fp: The path to the mapping file
-    :delimiter: The delimiter used in the mapping file
-    """
-    df = pd.read_csv(file_fp,
-                     sep=delimiter,
-                     header=[0, 1],
-                     skiprows=[2, 3, 4],
-                     na_filter=False)
-    df.replace('NA', nan, inplace=True)
-    # Get the tables in the dataframe while maintaining order
-    tables = []
-    for (table, header) in df.axes[1]:
-        tables.append(table)
-        for column in df[table]:
-            try:
-                df[table].assign(column=df[table][column].astype(fig.COLUMN_TYPES[table][column]))
-            except KeyError:
-                df[table].assign(column=df[table][column].astype('object'))
-    tables = list(dict.fromkeys(tables))
-    return tables, df
-
-
-def validate_mapping_file(file_fp, delimiter='\t'):
-    """
-    Checks the mapping file at file_fp for any errors.
-    Returns a list of the errors, an empty list means there
-    were no issues.
-    """
-    log('In validate_mapping_file')
-    errors = []
-    warnings = []
-    tables, df = load_mapping_file(file_fp, delimiter)
-
-    all_headers = []
-    study_name = None
-    # For each table
-    for table in tables:
-        # If the table shouldn't exist add and error and skip checking it
-        if table not in fig.TABLE_ORDER:
-            errors.append('-1\t-1\tTable Error: Table {} should not be the metadata'.format(table))
-            continue
-        table_df = df[table]
-        (new_errors, new_warnings, all_headers, study_name) = check_table(table_df, table, all_headers, study_name)
-        errors += new_errors
-        warnings += new_warnings
-
-    # Check for duplicate columns
-    dups = check_duplicate_cols(all_headers)
-    if dups:
-        for dup in dups:
-            locs = [i for i, header in enumerate(all_headers) if header == 'dup']
-            for loc in locs:
-                errors.append('1\t{}\tDuplicate Header Error: Duplicate header {}'.format(loc, dup))
-
-    # Check for missing tables
-    missing_tables = fig.METADATA_TABLES.difference(set(tables))
-    if missing_tables:
-        errors.append('-1\t-1\tMissing Table Error: Missing tables ' + ', '.join(missing_tables))
-
-    # Check for missing headers
-    missing_headers = REQUIRED_HEADERS.difference(set(all_headers))
-    if missing_headers:
-        errors.append('-1\t-1\tMissing Column Error: Missing required fields: ' + ', '.join(missing_headers))
-
-    return errors, warnings, study_name, df['Subjects']
-
-
 def is_numeric(s):
     """
     Check if the provided string is a number.
@@ -634,6 +319,8 @@ def is_numeric(s):
             return True
         except (TypeError, ValueError):
             pass
+    elif issubdtype(type(s), number):
+        return True
     return False
 
 
@@ -697,7 +384,7 @@ def generate_error_html(file_fp, errors, warnings):
     :errors: A list of the errors the metadata file produced
     :warnings: A list of the warnings the metadata file produced
     """
-    df = pd.read_csv(file_fp, sep='\t', header=[0, 1], skiprows=[2, 3, 4])
+    df = pd.read_csv(file_fp, sep='\t', header=[0, 1], skiprows=[2, 3, 4], na_filter=False)
     html = '<!DOCTYPE html>\n<html>\n'
     html += '<link type="text/javascript" rel="stylesheet" href="/CSS/stylesheet.css">\n'
     html += '<title> MMEDS Metadata Errors </title>\n'
@@ -748,8 +435,11 @@ def split_data(column):
     """
     result = defaultdict(list)
     if column.name == 'lat_lon':
-        for value in column:
+        log("name: {}, vals: {}".format(column.name, column))
+        # Skip the header
+        for value in column[1:]:
             parsed = value.strip('+').split('-')
+            log('Parsed: {}'.format(parsed))
             result['Latitude'].append(parsed[0])
             result['Longitude'].append(parsed[1])
     elif column.name == 'assembly_name':
@@ -760,40 +450,6 @@ def split_data(column):
     else:
         raise ValueError
     return result
-
-
-def load_MIxS_metadata(file_fp, skip_rows, unit_column):
-    """
-    A function for load and transforming the MIxS data in a pandas dataframe.
-    ========================================================================
-    :file_fp: The path to the file to convert
-    :skip_rows: The number of rows to skip after the header
-    :unit_column: A string. If None then the function checks each cell for units.
-    """
-    # Read in the data file
-    df = pd.read_csv(file_fp, header=0, sep='\t')
-    # Set the index to be the 'column_header' column
-    df.set_index('column_header', inplace=True)
-    # Remove rows with null indexes
-    df = df.loc[df.index.notnull()]
-    # Retrieve the unit column if one is specified
-    if unit_column is not None:
-        try:
-            units = df[unit_column]
-            df.drop(unit_column, axis=1, inplace=True)
-        except KeyError:
-            raise MetaDataError('The provided unit column is invalid.')
-    else:
-        units = {}
-    # Transpose the dataframe across the diagonal
-    df = df.T
-    # Drop unnamed columns
-    df.drop([x for x in df.axes[0] if 'Unnamed' in x], inplace=True)
-    # Drop any columns with only np.nan values
-    df.dropna(how='all', axis='columns', inplace=True)
-    # Replace np.nans with "NA"s
-    df.fillna('"NA"', inplace=True)
-    return df, units
 
 
 def MIxS_to_mmeds(file_fp, out_file, skip_rows=0, unit_column=None):
@@ -820,13 +476,13 @@ def MIxS_to_mmeds(file_fp, out_file, skip_rows=0, unit_column=None):
             first = df[item][0].split(' ')
             # If the value is numeric grab the units in the data cell
             if is_numeric(first[0]):
-                unit_col = item  # + ' ({})'.format(' '.join(first[1:]))
+                unit_col = item
                 df[item] = df[item].map(lambda x: x.split(' ')[0])
             else:
                 unit_col = item
         # Add the units to the header if available
         else:
-            unit_col = item  # + ' ({})'.format(units[item])
+            unit_col = item
         fig.MIXS_MAP[('AdditionalMetaData', str(unit_col))] = str(unit_col)
         fig.MMEDS_MAP[item] = ('AdditionalMetaData', str(unit_col))
         all_cols['AdditionalMetaData'].append(str(unit_col))
@@ -836,6 +492,7 @@ def MIxS_to_mmeds(file_fp, out_file, skip_rows=0, unit_column=None):
     for col in df.columns:
         (table, column) = fig.MMEDS_MAP[col]
         if ':' in column:
+            log('Table: {}, Column: {}'.format(table, column))
             cols = column.split(':')
             data = split_data(df[col])
             for new_col in cols:
@@ -875,7 +532,7 @@ def write_mmeds_metadata(out_file, meta, all_cols, num_rows):
             column_type.append(str(md_template[table][column].iloc[0]))
             column_unit.append(str(md_template[table][column].iloc[1]))
             column_required.append(str(md_template[table][column].iloc[2]))
-        except KeyError:
+        except (KeyError, IndexError):
             column_type.append('')
             column_unit.append('')
             column_required.append('')
@@ -892,10 +549,11 @@ def write_mmeds_metadata(out_file, meta, all_cols, num_rows):
             for table, column in zip(table_row, column_row):
                 # Add the value to the row
                 try:
-                    row.append(meta[(table, column)][i])
-                # If a value doesn't exist for this table,column insert NA
-                except KeyError:
-                    row.append('"NA"')
+                    row.append(meta[(table, column)][i].strip('"'))
+                # If a value doesn't exist for this table, column insert NA
+                except (KeyError, IndexError):
+                    log('Key error for {}, {}, {}'.format(table, column, i))
+                    row.append('NA')
             f.write('\t'.join(row) + '\n')
 
 
@@ -938,10 +596,6 @@ def send_email(toaddr, user, message='upload', testing=False, **kwargs):
     :kwargs: Any information that is specific to a paricular message type
     """
     log('Send email to: {} on behalf of {}'.format(toaddr, user))
-    if testing:
-        return
-    for key in kwargs.keys():
-        log('{}: {}'.format(key, kwargs[key]))
 
     # Templates for the different emails mmeds sends
     if message == 'upload':
@@ -986,14 +640,16 @@ def send_email(toaddr, user, message='upload', testing=False, **kwargs):
         script += ' -A {summary}'.format(kwargs['summary'])
     cmd = script.format(body=email_body, subject=subject, toaddr=toaddr)
     log(cmd)
-    run(['/bin/bash', '-c', cmd], check=True)
+    if not testing:
+        run(['/bin/bash', '-c', cmd], check=True)
+    return cmd
 
 
 def pyformat_translate(value):
     """ Convert from numpy to standard python datatypes. """
-    if isinstance(value, np.int64):
+    if isinstance(value, int64):
         result = int(value)
-    elif isinstance(value, np.float64):
+    elif isinstance(value, float64):
         result = float(value)
     else:
         result = value
@@ -1007,17 +663,38 @@ def setup_environment(module):
     :module: A string. The name of the module to load.
     """
     # Check there is nothing in module that could cause problems
-    if not module.replace('_', '').replace('-', '').isalnum():
-        raise InvalidModuleError('{} is not a valid module name.' +
-                                 'Modules may only contain letters, numbers, "_", and "-"')
+    if not module.replace('/', '').replace('_', '').replace('-', '').replace('.', '').isalnum():
+        raise InvalidModuleError('{} is not a valid module name. '.format(module) +
+                                 'Modules may only contain letters, numbers, "/", "_", "-", and "."')
 
-    log('Setup environment for {}'.format(module))
-    run(['/bin/bash', '-c', 'module use ~/.modules/modulefiles'], check=True)
+    module_file = (fig.MODULE_ROOT / module).read_text()
     new_env = environ.copy()
-    output = run(['/bin/bash', '-c', 'module load {}; echo $PATH;'.format(module)],
-                 capture_output=True, env=new_env, check=True)
-    new_env['PATH'] = output.stdout.decode('utf-8').strip()
-    log('New path: {}'.format(new_env['PATH']))
+    variables = {}
+
+    for line in module_file.splitlines():
+        if line.startswith('#'):
+            continue
+        # Update any values with local variables
+        if '$' in line:
+            for variable, path in variables.items():
+                line = line.replace('${}'.format(variable), path)
+        if '~' in line:
+            for variable, path in variables.items():
+                line = line.replace('~', new_env['HOME'])
+        parts = line.strip().split(' ')
+        # Set locally used variables
+        if parts[0] == 'set':
+            variables[parts[1]] = parts[2]
+        # Add to PATH
+        elif parts[0] == 'prepend-path':
+            new_env[parts[1]] = '{}:{}'.format(parts[2], new_env[parts[1]])
+        # Remove from PATH
+        elif parts[0] == 'remove-path':
+            path_parts = new_env[parts[1]].split(':')
+            new_env[parts[1]] = ':'.join([part for part in path_parts if not part == parts[1]])
+        # Set environment variables
+        elif parts[0] == 'setenv':
+            new_env[parts[1]] = parts[2]
     return new_env
