+++ sql/tables.sql
@@ -1,5 +1,5 @@
 -- MySQL Script generated by MySQL Workbench
--- Tue 26 Feb 2019 08:14:31 AM EST
+-- Thu 07 Mar 2019 10:56:46 AM EST
 -- Model: New Model    Version: 1.0
 -- MySQL Workbench Forward Engineering
 
@@ -81,6 +81,7 @@ DROP TABLE IF EXISTS `mmeds_data1`.`Type` ;
 CREATE TABLE IF NOT EXISTS `mmeds_data1`.`Type` (
   `idType` INT NOT NULL,
   `SpecimenType` VARCHAR(45) NULL DEFAULT NULL,
+  `UberonCodeType` VARCHAR(45) NULL,
   PRIMARY KEY (`idType`))
 ENGINE = InnoDB;
 
@@ -114,6 +115,7 @@ DROP TABLE IF EXISTS `mmeds_data1`.`BodySite` ;
 CREATE TABLE IF NOT EXISTS `mmeds_data1`.`BodySite` (
   `idBodySite` INT NOT NULL,
   `SpecimenBodySite` VARCHAR(45) NULL DEFAULT NULL,
+  `UberonCodeBodySite` VARCHAR(45) NULL,
   PRIMARY KEY (`idBodySite`))
 ENGINE = InnoDB;
