+++ travis_setup.sh
@@ -62,9 +62,14 @@ ln -sf ~/.modules/qiime2 ~/miniconda2/envs/qiime2;
 ln -sf ~/.modules/mmeds-stable ~/miniconda2/envs/mmeds-stable;
 
 # Create links to the module files
+# TEMPORARY
+rm -rf ~/.modules/modulefiles/qiime1
+rm -rf ~/.modules/modulefiles/qiime2
+mkdir ~/.modules/modulefiles/qiime
+mkdir ~/.modules/modulefiles/qiime2
 ln -sf $REPO_DIR/modules/mmeds-stable ~/.modules/modulefiles/mmeds-stable
-ln -sf $REPO_DIR/modules/qiime2 ~/.modules/modulefiles/qiime2
-ln -sf $REPO_DIR/modules/qiime1 ~/.modules/modulefiles/qiime1
+ln -sf $REPO_DIR/modules/qiime2 ~/.modules/modulefiles/qiime2/2019.1
+ln -sf $REPO_DIR/modules/qiime1 ~/.modules/modulefiles/qiime/1.9.1
 
 cd $REPO_DIR;
 echo "Finished setup. In $(pwd)"
